refactor: split session runtime responsibilities
