+++ packages/control-plane/src/session/runtime-contracts.ts
@@ -1,0 +1,9 @@
+export interface SessionRuntimeClient {
+  participantId: string;
+  userId: string;
+  lastFetchHistoryAt?: number;
+}
+
+export type SessionConnectionKind =
+  | { kind: "sandbox"; sandboxId?: string }
+  | { kind: "client"; wsId?: string };
